// src/types/colors/colors.ts
export enum CodeColor {
    CONTAINER_DARK = '#2c2b2b',
    CONTAINER_LIGHT = '#fff',

    FONDO_DARK = '#2c2b2b',
    FONDO_LIGHT = '#e6eaed',

    TEXT_DARK = '#fff',
    TEXT_LIGHT = '#000',

    FOOTER_DARK = '#1A1C1E',
    FOOTER_LIGHT = '#fff',

    BACKDROP_DARK = '#1A1C1E',
    BACKDROP_LIGHT = '#fff',

    ROW_TABLE_DARK = '#2c2b2b',
    ROW_TABLE_LIGHT = '#f4f4f4',

    HEADER_DARK = '#1A1C1E',
    HEADER_LIGHT = '#fff',
  }
  